refactor: extract c2-config crate for unified configuration
Create a new c2-config leaf crate (zero external deps) containing all
configuration structs previously scattered across the workspace:

- IpcConfig: merged server-side (18 fields) and client-side (5 fields)
  into a single unified struct (20 fields) with consistent types
  (u64 for sizes, f64 for durations, u32 for counts)
- PoolConfig: moved from c2-mem
- RelayConfig: moved from c2-relay

Key changes:
- reply_chunk_size (server) and chunk_size (client) merged into
  single chunk_size field (default 128KB)
- All downstream crates re-export from c2-config for backward compat
- Unified IpcConfig includes validate() with 6 constraint checks
- Added docs/issues/chunk-management-gaps.md documenting unused
  chunk GC config fields and memory leak risks

+++ src/c_two/_native/c2-ipc/src/sync_client.rs
@@ -82,7 +82,7 @@ impl SyncClient {
 
     /// Whether the client has a SHM pool and data exceeds the threshold.
     pub fn should_use_shm(&self, data_len: usize) -> bool {
-        self.inner.pool.is_some() && data_len > self.inner.config.shm_threshold
+        self.inner.pool.is_some() && data_len > self.inner.config.shm_threshold as usize
     }
 
     /// Allocate from the client SHM pool and write data in a single lock scope.
